requirements.txt
click
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
jinja2>=2.11.3
MarkupSafe==1.1.1
numpy
Werkzeug==1.0.1
tensorflow-cpu==2.7.0
tensorflow_text
flask-cors
Flask-RESTful